refactor: more small file refactors
