+++ router/src/creator/mod.rs
@@ -1,9 +1,11 @@
 // Core traits and types
 pub mod queue;
 pub mod traits;
+pub mod types;
 
-pub use queue::{CreatorConfig, SimpleTaskQueue, TaskQueue};
+pub use queue::{SimpleTaskQueue, TaskQueue};
 pub use traits::Creator;
+pub use types::CreatorConfig;
 
 // Test module
 #[cfg(test)]
