+++ router/src/creator/types.rs
@@ -1,0 +1,14 @@
+#[derive(Debug, Clone)]
+pub struct CreatorConfig {
+    pub polling_interval_ms: u64,
+    pub timeout_ms: u64,
+}
+
+impl Default for CreatorConfig {
+    fn default() -> Self {
+        Self {
+            polling_interval_ms: 100,
+            timeout_ms: 5000,
+        }
+    }
+}
