+++ router/src/executor/traits.rs
@@ -1,6 +1,8 @@
 use anyhow::Result;
 use async_trait::async_trait;
 
+use crate::executor::types::ExecutionResult;
+
 #[async_trait]
 pub trait VerificationExecutor<T = (), V = ()>: Send + Sync
 where
@@ -13,13 +15,4 @@ where
         verification_data: V,
         task_data: Option<&T>,
     ) -> Result<ExecutionResult>;
-}
-
-#[derive(Debug, Clone)]
-pub struct ExecutionResult {
-    pub transaction_hash: String,
-    pub block_number: Option<u64>,
-    pub gas_used: Option<u64>,
-    pub status: Option<bool>,
-    pub contract_address: Option<String>,
 }
