+++ router/src/executor/types.rs
@@ -24,3 +24,12 @@ impl VerificationData {
         self
     }
 }
+
+#[derive(Debug, Clone)]
+pub struct ExecutionResult {
+    pub transaction_hash: String,
+    pub block_number: Option<u64>,
+    pub gas_used: Option<u64>,
+    pub status: Option<bool>,
+    pub contract_address: Option<String>,
+}
