+++ router/src/ingress/mod.rs
@@ -1,2 +1,2 @@
-pub mod http_server;
+pub mod server;
 pub mod types;
