+++ usecase-counter/router/src/factories.rs
@@ -9,8 +9,8 @@ use alloy_signer_local::PrivateKeySigner;
 use anyhow::Result;
 use commonware_avs_bindings::WalletProvider;
 use commonware_avs_bindings::counter::Counter;
-use commonware_avs_router::creator::queue::{CreatorConfig, SimpleTaskQueue};
-use commonware_avs_router::ingress::http_server::start_http_server;
+use commonware_avs_router::creator::{CreatorConfig, SimpleTaskQueue};
+use commonware_avs_router::ingress::server::start_http_server;
 use commonware_eigenlayer::config::AvsDeployment;
 use std::{env, str::FromStr};
